create file to migrate menu data
